🚚 [RUMF-1327] move function to read synthetics worker values in core
