+++ packages/core/src/index.ts
@@ -72,3 +72,8 @@ export { createContextManager } from './tools/contextManager'
 export { limitModification } from './tools/limitModification'
 export { ContextHistory, ContextHistoryEntry, CLEAR_OLD_CONTEXTS_INTERVAL } from './tools/contextHistory'
 export { SESSION_COOKIE_NAME } from './domain/session/sessionCookieStore'
+export {
+  willSyntheticsInjectRum,
+  getSyntheticsTestId,
+  getSyntheticsResultId,
+} from './domain/synthetics/syntheticsWorkerValues'
