+++ packages/rum-core/src/boot/rumPublicApi.spec.ts
@@ -1,13 +1,9 @@
 import type { RelativeTime, TimeStamp } from '@datadog/browser-core'
 import { ONE_SECOND, getTimeStamp, display, DefaultPrivacyLevel } from '@datadog/browser-core'
+import { cleanupSyntheticsWorkerValues, mockSyntheticsWorkerValues } from '../../../core/test/syntheticsWorkerValues'
 import { initEventBridgeStub, deleteEventBridgeStub } from '../../../core/test/specHelper'
 import type { TestSetupBuilder } from '../../test/specHelper'
-import {
-  cleanupSyntheticsWorkerValues,
-  mockSyntheticsWorkerValues,
-  noopRecorderApi,
-  setup,
-} from '../../test/specHelper'
+import { noopRecorderApi, setup } from '../../test/specHelper'
 import type { HybridInitConfiguration, RumInitConfiguration } from '../domain/configuration'
 import { ActionType } from '../rawRumEvent.types'
 import type { RumPublicApi, StartRum, RecorderApi } from './rumPublicApi'
