+++ packages/rum-core/src/domain/contexts/syntheticsContext.spec.ts
@@ -1,5 +1,5 @@
-import { mockSyntheticsWorkerValues, cleanupSyntheticsWorkerValues } from '../../../test/specHelper'
-import { getSyntheticsContext, willSyntheticsInjectRum } from './syntheticsContext'
+import { cleanupSyntheticsWorkerValues, mockSyntheticsWorkerValues } from '../../../../core/test/syntheticsWorkerValues'
+import { getSyntheticsContext } from './syntheticsContext'
 
 describe('getSyntheticsContext', () => {
   afterEach(() => {
@@ -48,35 +48,5 @@ describe('getSyntheticsContext', () => {
     mockSyntheticsWorkerValues({ publicId: 'foo' }, 'cookies')
 
     expect(getSyntheticsContext()).toBeUndefined()
-  })
-})
-
-describe('willSyntheticsInjectRum', () => {
-  afterEach(() => {
-    cleanupSyntheticsWorkerValues()
-  })
-
-  it('returns false if nothing is defined', () => {
-    mockSyntheticsWorkerValues({}, 'globals')
-
-    expect(willSyntheticsInjectRum()).toBeFalse()
-  })
-
-  it('returns false if the INJECTS_RUM global variable is false', () => {
-    mockSyntheticsWorkerValues({ injectsRum: false }, 'globals')
-
-    expect(willSyntheticsInjectRum()).toBeFalse()
-  })
-
-  it('returns true if the INJECTS_RUM global variable is truthy', () => {
-    mockSyntheticsWorkerValues({ injectsRum: true }, 'globals')
-
-    expect(willSyntheticsInjectRum()).toBeTrue()
-  })
-
-  it('returns true if the INJECTS_RUM cookie is truthy', () => {
-    mockSyntheticsWorkerValues({ injectsRum: true }, 'cookies')
-
-    expect(willSyntheticsInjectRum()).toBeTrue()
   })
 })
